third_party/blink/web_tests/platform/generic/fast/canvas/OffscreenCanvas-transferToImageBitmap-state-persistency-expected.txt

PASS imgData[0] is 0
PASS imgData[1] is 255
PASS imgData[2] is 0
PASS imgData[3] is 255
PASS imgData[0] is 0
PASS imgData[1] is 0
PASS imgData[2] is 0
PASS imgData[3] is 0
PASS successfullyParsed is true

TEST COMPLETE
